examples/echo/pyproject.toml
[project]
name = "echo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.0"
dependencies = ["asyncord", "python-dotenv>=1.0.1"]

[tool.uv.sources]
asyncord = { workspace = true }
